db stat

+++ src/category.h
@@ -1,8 +1,8 @@
 #include "server.h"
 #include "rio.h"
 
-ssize_t categoryObjectSize(robj *o);
+long long categoryObjectSize(robj *o);
 void doCalculateCategory();
 void ccCommand(client *c);
-void addCateforyStats(robj *key, int size, dict* d);
+void addCateforyStats(robj *key, long long size, dict* d);
